Studio: rename compact-fields tests to responsive-fields
