+++ cypress/fixtures/studio-responsive-fields-code.json
@@ -1,0 +1,3 @@
+{
+  "code": "<style> \n    *, *::before, *::after { box-sizing: border-box; margin: 0; padding: 0; }\n    body {\n      font-family: Mulish Arial, Helvetica, sans-serif;\n      background: #F9FAFB;\n    }\n    .wrapper {\n      background: #FFFFFF;\n      color: #3F51B5;\n      padding: 40px;\n      margin: 20px 0;\n    }\n    .card {\n      width: 300px;\n      background: #ffffff;\n      border-radius: 8px;\n      box-shadow: 0 2px 16px rgba(0,0,0,0.12);\n      overflow: hidden;\n    }\n    .card__image {\n      width: 100%;\n      display: block;\n    }\n    .card__body {\n      padding: 20px 20px 28px;\n    }\n    .card__date {\n      display: inline-block;\n      font-family: 'Courier New', monospace;\n      font-size: 13px;\n      color: #1a1a1a;\n      padding: 2px 6px;\n      margin-bottom: 14px;\n      color: #1a1a1a;\n    }\n    .card__title {\n      font-family: Arial, Helvetica, sans-serif;\n      font-size: 28px;\n      font-weight: 700;\n      line-height: 1.2;\n      color: #111111;\n      letter-spacing: -0.3px;\n    }\n  </style>\n<div class=\"wrapper\" data-cy=\"multi_line_text_field\" ><p style=\"color: #3F51B5; font-size: 16px;\">{{this.multi_line_text_field}}</p></div>\n<div class=\"wrapper\" data-cy=\"wysiwyg_field\">{{this.wysiwyg_field}}</div>\n<div class=\"wrapper\" data-cy=\"markdown_field\">{{this.markdown_field}}</div>\n<div class=\"wrapper\" data-cy=\"media_field\" >{{this.media_field.getImage()}}</div>\n<div style=\"color: #3F51B5; padding: 40px; background: #F1F1F4;\">\n <article class=\"card\">\n    <img\n      class=\"card__image\"\n      src=\"{{this.media_field.getImage()}}\"\n      alt=\"A boy and a robot playing Battleship while browser logos watch\"\n\t  style=\"height:250px;width: 100%;\"\n    />\n    <div class=\"card__body\">\n      <p class=\"card__date\" data-cy=\"date___time_field\">{{this.date___time_field}}</time>\n      <h1 class=\"card__title\">Optimizing Organization Websites for&nbsp;GEO</h1>\n    </div>\n  </article>\n</div>"
+}
